Refactoring: Anno tools.
