+++ src/index.html
@@ -113,7 +113,7 @@
                                     </a>
     							</div>
     							<div class="btn-group" role="group">
-                                    <a class="btn btn-default js-tool-btn" href="#" data-type="rect" title="rectangle">
+                                    <a class="btn btn-default js-tool-btn-rect" href="#" data-type="rect" title="rectangle">
                                         <i class="fa fa-square-o fa-2x"></i>
                                     </a>
                                 </div>
