+++ src/page/pdf/index.js
@@ -6,6 +6,7 @@
 
 
  export default {
+    // TODO May need to fix..
     loadFiles,
     globalVariable
  };
